database -> articles.prix : From INT To FLOAT

+++ backend/database/schema.sql
@@ -24,7 +24,7 @@ CREATE TABLE IF NOT EXISTS `le_comptoir_des_seelies`.`articles` (
   `id` INT NOT NULL AUTO_INCREMENT,
   `nom` VARCHAR(45) NOT NULL,
   `image` VARCHAR(45) NOT NULL,
-  `prix` INT NOT NULL,
+  `prix` FLOAT NOT NULL,
   `ajout_date` DATETIME NOT NULL,
   `nb_ventes` INT NULL DEFAULT NULL,
   `taille` INT NOT NULL,
